+++ dglib_api/src/main/java/com/dglib/repository/NoticeImageRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dglib.entity.NoticeImage;
 
-public interface NoticeImageRepository extends JpaRepository<NoticeImage, Integer>{
+public interface NoticeImageRepository extends JpaRepository<NoticeImage, Long>{
 	
-	List<NoticeImage> findByNotice_Ano(int ano);
+	List<NoticeImage> findByNotice_Ano(Long ano);
 
 }
